# Optimize and improve the proc_macro RPC interface for cross-thread execution