--- a/compiler/rustc_expand/src/proc_macro_server.rs
+++ b/compiler/rustc_expand/src/proc_macro_server.rs
@@ -388,7 +388,10 @@ impl<'a> Rustc<'a> {
     }
 
     fn lit(&mut self, kind: token::LitKind, symbol: Symbol, suffix: Option<Symbol>) -> Literal {
-        Literal { lit: token::Lit::new(kind, symbol, suffix), span: server::Span::call_site(self) }
+        Literal {
+            lit: token::Lit::new(kind, symbol, suffix),
+            span: server::Context::call_site(self),
+        }
     }
 }
 
@@ -484,7 +487,7 @@ impl server::Group for Rustc<'_> {
         Group {
             delimiter,
             stream,
-            span: DelimSpan::from_single(server::Span::call_site(self)),
+            span: DelimSpan::from_single(server::Context::call_site(self)),
             flatten: false,
         }
     }
@@ -510,7 +513,7 @@ impl server::Group for Rustc<'_> {
 
 impl server::Punct for Rustc<'_> {
     fn new(&mut self, ch: char, spacing: Spacing) -> Self::Punct {
-        Punct::new(ch, spacing == Spacing::Joint, server::Span::call_site(self))
+        Punct::new(ch, spacing == Spacing::Joint, server::Context::call_site(self))
     }
     fn as_char(&mut self, punct: Self::Punct) -> char {
         punct.ch
@@ -712,15 +715,6 @@ impl server::Span for Rustc<'_> {
             format!("{:?} bytes({}..{})", span.ctxt(), span.lo().0, span.hi().0)
         }
     }
-    fn def_site(&mut self) -> Self::Span {
-        self.def_site
-    }
-    fn call_site(&mut self) -> Self::Span {
-        self.call_site
-    }
-    fn mixed_site(&mut self) -> Self::Span {
-        self.mixed_site
-    }
     fn source_file(&mut self, span: Self::Span) -> Self::SourceFile {
         self.sess.source_map().lookup_char_pos(span.lo()).file
     }
@@ -798,6 +792,18 @@ impl server::Span for Rustc<'_> {
             // our current call site.
             raw_span.with_def_site_ctxt(expn_id)
         })
+    }
+}
+
+impl server::Context for Rustc<'_> {
+    fn def_site(&mut self) -> Self::Span {
+        self.def_site
+    }
+    fn call_site(&mut self) -> Self::Span {
+        self.call_site
+    }
+    fn mixed_site(&mut self) -> Self::Span {
+        self.mixed_site
     }
 }