… Cargo.lock
… compiler/rustc_expand/Cargo.toml
… library/proc_macro/src/bridge/handle.rs
… library/proc_macro/src/bridge/mod.rs
… library/proc_macro/src/bridge/rpc.rs
… library/proc_macro/src/bridge/server.rs
… library/proc_macro/src/lib.rs
… src/tools/tidy/src/deps.rs