This series of patches performs various changes to the RPC interface used by `proc_macro` to improve its performance. From my local testing (which involved doing `check --test` runs on serde's test suite, based on https://github.com/rust-lang/rust/pull/49219#issuecomment-439648025) these pages have small improvements to same-thread execution, and significant improvements to cross-thread execution.

Some of the major changes include:
* Stash frequently accessed interned values (`Span::{def,call,mixed}_site`) in the client bridge, so no RPC is required to access them,
* Support non-blocking messaging to the server for calls which can be handled async (such as `TokenStream::clone()` or composing `Token{Tree,Stream}`s together),
* Reduce the number of messages required for common `TokenStream` operations (removes `TokenStreamBuilder`, `TokenStreamIter` in favour of passing `Vec<TokenTree>` over RPC),
* Store an `Option<bridge::client::TokenStream>` in the `TokenStream` type to avoid `TokenStream::new()` call overhead,
* Add a local symbol interner within the client bridge to hold `Symbol`s in the client, along with support to serialize them over RPC,
* Decompose the token types `Group`, `Punct`, `Ident`, and `Literal` into struct types around more primitive handles (`TokenStream`, `Symbol`, and `Span`), meaning that common operations on these types, such as accessing `.span()` or `.delimiter()` require no RPC calls,
* Use `crossbeam-channel` for cross-thread message passing, to avoid `mpsc::channel`'s overhead,
* Added a flag (`-Z proc_macro_cross_thread`) to turn on the `CrossThread` execution strategy for proc macros without rebuilding.

The PR is organized into a sequence of commits which build on top of one another, and could be reviewed independently.

Some example times from my machine (there's unfortunately variation in times run-to-run, but I tried to pick representative examples):

<details>
<summary>PR (SameThread)</summary>

```sh
~/src/serde/test_suite$ ( export RUSTFLAGS="-Z proc_macro_cross_thread=no"; cargo +dev check --tests && rm -f ../target/debug/deps/libtest_*.rmeta && time cargo +dev check --tests )
    Finished dev [unoptimized + debuginfo] target(s) in 0.02s
    Checking serde_test_suite v0.0.0 (/home/nika/src/serde/test_suite)
    Finished dev [unoptimized + debuginfo] target(s) in 0.76s

real	0m0.779s
user	0m2.256s
sys	0m0.459s
```
</details>

<details>
<summary>PR (CrossThread)</summary>

```sh
~/src/serde/test_suite$ ( export RUSTFLAGS="-Z proc_macro_cross_thread=yes"; cargo +dev check --tests && rm -f ../target/debug/deps/libtest_*.rmeta && time cargo +dev check --tests )
    Finished dev [unoptimized + debuginfo] target(s) in 0.02s
    Checking serde_test_suite v0.0.0 (/home/nika/src/serde/test_suite)
    Finished dev [unoptimized + debuginfo] target(s) in 0.83s

real	0m0.847s
user	0m2.971s
sys	0m0.579s
```
</details>

<details>
<summary>HEAD (SameThread)</summary>

```sh
~/src/serde/test_suite$ cargo +dev check --tests && rm -f ../target/debug/deps/libtest_*.rmeta && time cargo +dev check --tests
    Finished dev [unoptimized + debuginfo] target(s) in 0.02s
    Checking serde_test_suite v0.0.0 (/home/nika/src/serde/test_suite)
    Finished dev [unoptimized + debuginfo] target(s) in 0.85s

real	0m0.866s
user	0m2.608s
sys	0m0.395s
```
</details>

<details>
<summary>HEAD (CrossThread2)</summary>

```sh
~/src/serde/test_suite$ cargo +dev check --tests && rm -f ../target/debug/deps/libtest_*.rmeta && time cargo +dev check --tests
    Finished dev [unoptimized + debuginfo] target(s) in 0.02s
    Checking serde_test_suite v0.0.0 (/home/nika/src/serde/test_suite)
    Finished dev [unoptimized + debuginfo] target(s) in 7.49s

real	0m7.504s
user	0m8.036s
sys	0m13.600s
```
</details>

(note: there will likely be changes required to RA to support these changes as well, however this patch currently includes no changes to RA or it's copy of the `proc_macro` crate)