@eddyb 2021-07-02
Just to get ahead of this a bit: I was aware that some of this was being worked on by @mystor, and long-term I believe we want all of these changes, but I'm not sure how to approach this PR with all of them at once.

If I find enough free time, I could maybe help split into several parts (there's some interesting design changes that aren't strictly tied to cross-thread-specific optimizations, like drastically cutting down on RPC "object" types, keeping interned symbols a strings in the client, etc.)

But as of right now I can't promise anything. If someone else is willing to review this in bulk, feel free to go ahead, but I doubt we could land it without an MCP.