@mystor 2021-07-02
I'd be willing to split this into smaller PRs based on the individual patches in the review (each of which should be independent). I unfortunately don't know how to do this on github, however (reading suggests it's not possible). Perhaps we could start with a smaller PR which only includes the changes up to https://github.com/rust-lang/rust/pull/86816/commits/afc36cc1af9dbb18a7b6d10b7f8af95e081beff1, then do the range for the `wait`/`nowait` changes (up to https://github.com/rust-lang/rust/pull/86816/commits/8db4f28c53176ef190b7cba1fdc2b95d66af030d), and then the non-remote object changes (up to https://github.com/rust-lang/rust/pull/86816/commits/efedf62444cde5fb89febabff0465deb0ac4b9c0). I think it should also be possible to rebase the struct changes to not include the wait/nowait changes if they're particularly controversial (as they only impact `CrossThread`).