@mystor 2021-07-02
Closing this PR for now in favour of #86822 which should provide the `SameThread` benefits as this patch, but is smaller. If that patch is merged, the `CrossThread` changes in this PR could be rebased on top of those changes.